Updates lager_syslog to not crash on syslog crash
When syslog crashes on bad messages (the port crashing specifically),
lager_syslog_backend loses the current log level and reverts back to
the initial arguments. For those that set log level at run time, this
setting is lost and needs to be manually reset.

Instead, catch errors from syslog so state is preserved and restart
logging.