KZOO-366: optionally write dialyzer output to a file
This adds a new option to `check-dialyzer.escript` to optionally write the
Dialyzer output to a file while still print in in standard IO.

`dialyzer.mk` is set as such to pass the `DIALYZER_OUTPUT` env variable
as `output-file` to `check-dialyzer.escript` if set.

When `CIRCLECI` env var is set, the default `DIALYZER_OUTPUT` is forced
with the value: `${PWD}/dialyzer_output.log`.

Dialyzer has an `output_file` option but this only works when Dialyzer
if invoked from CLI. Calling `dialyzer:run/1` ALWAYS return's the warnings
without printing it to screen or writing it to `output_file`.

So instead of setting that option, we do the printing and writing to output
directly in `check-dialyzer.escript`.