KZOO-366: fix a condition when no erlang file to dialyze in dialyzer-changed
Follow up the changes in https://github.com/2600hz/kazoo5/pull/96 to
save Dialyzer output saved into a file, there is a specific edge case
when there is no `.erl`/`.beam` file to Dialyze in `CHANGED` variable.
In this case `check-dialyzer.escript` will not create and save the
result into the output file, and this result the CI step unable to save
the dialyze output file as artifact as a proof that dialyzer was ran.

This changes the script to always output something the file to report
even if there were no file passed to the script.

Also this fixes another edge case, `--bulk` option from
`dialyze-it-changed` won't be passed to `dialyze-it` because of using
`if` clause in the `dialyze-it-changed` step calling `make dialze-it`
without re-exporting `CHECK_DIALYZER_OPTS`.