tinkering with rebar3 as alternative experimental build system (#117)
This introduce rebar3 as an alternative build system, only for use
locally in your dev env.

Current Makefile system is robust and gets the job done. But it
increasingly getting difficult to run this in different env and OS,
and it is ridiculously SLOW. Average time is something around
7mins. If you run in VM it may be more close to 10min. If you add
macOS in equation, it is infinity (something around 25-3m min).

With Rebar3 the build time is something under or around 2min, for
everything, including deps.

Currently this is just a simple configuration to test out
rebar. `xref`, `dialyzer`, test and other targets are not tested, and
are not properly configured.

This tries to not depend on calling Makefile targets as much as
possible. This reduce the problems on different platforms especially
windows and macos. Some scripts was added to replace some tasks needed
to setup files for compilation.

you must run eunit with KAZOO_CONFIG=rel/config-test.ini rebar3 eunit