sync rebar3 deps with make/deps.mk + add comment which branch we are using
This sync rebar.config with latest updates to make/deps.mk.

Also both files are now strictly referencing a commit or tag instead of
branch name, so deps are always get resolved exactly as we want and
avoiding stale deps in build folder.

Both files are now mentioning which branch their commit is referencing so
it would be easier for us to check out the code and maintain any of our
patches.