KZOO-587: ignore apps that can't be loaded (#191)
When building service pack branches, sometimes the core fix branch
won't have newer code that other KAZOO apps on the version branch will
reference.

For instance, on 5.4, kazoo_monitor is present and some apps will
expect it to be included in the TAGS file.

However, the fix-5.4.7 was tagged before the kazoo_monitor work was
merged, so all subsequent SP builds based off that fix branch will
fail to compile.

This commit should skip those apps that can't be loaded